requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
